add xxh3 hash, fix queries

+++ src/common/crc32.c
@@ -79,10 +79,10 @@ unsigned int xcrc32 (const unsigned char *buf, int len, unsigned int init)
   return crc;
 }
 
-uint32_t crc32(const char *buf, uint32_t len, uint32_t init) {
+uint64_t crc32(const char *buf, uint32_t len, uint32_t init) {
 	return xcrc32((const unsigned char*)buf, (int)len, (unsigned int)init);
 }
 
-uint32_t crc32_get(const char *buf) {
+uint64_t crc32_get(const char *buf) {
 	return xcrc32((const unsigned char*)buf, strlen(buf), 0);
 }
