+++ src/common/murmurhash.c
@@ -4,7 +4,7 @@
 #include <string.h>
 #include "murmurhash.h"
 
-uint32_t murmurhash (const char *key, uint32_t len, uint32_t seed) {
+uint64_t murmurhash (const char *key, uint32_t len, uint32_t seed) {
   uint32_t c1 = 0xcc9e2d51;
   uint32_t c2 = 0x1b873593;
   uint32_t r1 = 15;
@@ -61,6 +61,6 @@ uint32_t murmurhash (const char *key, uint32_t len, uint32_t seed) {
   return h;
 }
 
-uint32_t murmurhash_get (const char *key) {
+uint64_t murmurhash_get (const char *key) {
 	return murmurhash (key, strlen(key), 0);
 }
