+++ src/common/murmurhash.h
@@ -1,4 +1,4 @@
 #define MURMURHASH_VERSION "0.0.3"
 
-uint32_t murmurhash (const char *, uint32_t, uint32_t);
-uint32_t murmurhash_get (const char *key);
+uint64_t murmurhash (const char *, uint32_t, uint32_t);
+uint64_t murmurhash_get (const char *key);
