+++ src/dstructures/ht.c
@@ -138,10 +138,10 @@ void alligator_ht_forfree(alligator_ht *h, void *funcfree)
 	pthread_rwlock_unlock(&h->rwlock);
 }
 
-uint32_t alligator_ht_strhash(const char *buf, uint32_t len, uint32_t initial) {
+uint64_t alligator_ht_strhash(const char *buf, uint32_t len, uint32_t initial) {
 	return tommy_strhash_u32(initial, buf);
 }
 
-uint32_t alligator_ht_strhash_get(const char *buf) {
+uint64_t alligator_ht_strhash_get(const char *buf) {
 	return tommy_strhash_u32(0, buf);
 }
