+++ src/metric/metrictree.h
@@ -27,7 +27,7 @@ typedef struct sort_order_node {
 typedef struct sortplan
 {
 	char *plan[65535];
-	uint32_t hash[65535];
+	uint64_t hash[65535];
 	uint64_t len[65535];
 	uint8_t is_collission[65535];
 	alligator_ht* check_collissions;
@@ -38,10 +38,10 @@ typedef struct labels_t
 {
 	char *name;
 	size_t name_len;
-	uint32_t name_hash;
+	uint64_t name_hash;
 	char *key;
 	size_t key_len;
-	uint32_t key_hash;
+	uint64_t key_hash;
 	sortplan *sort_plan;
 	uint8_t allocatedname;
 	uint8_t allocatedkey;
